@@ -19,7 +19,9 @@ import { ExerciseSix } from './exercise-six';
 import { ExerciseSeven } from './exercise-seven';
 import { ExerciseEight } from './exercise-eight';
 
-
+import { ExerciseNine } from './exercise-nine'
+import { ExerciseTen } from './exercise-ten'
+import { ExerciseEleven } from './exercise-eleven'
 
 export const App: React.FC = () => {
   return (
@@ -43,6 +45,9 @@ export const App: React.FC = () => {
       <Link to="/exercise-six">Exercise six</Link>
       <Link to="/exercise-seven">Exercise seven</Link>
       <Link to="/exercise-eight">Exercise eight</Link>
+      <Link to="/exercise-nine">Exercise nine</Link>
+      <Link to="/exercise-ten">Exercise ten</Link>
+      <Link to="/exercise-eleven">Exercise eleven</Link>
 
       <Route path="/exercise-one">
         <ExerciseOne />
@@ -90,17 +95,23 @@ export const App: React.FC = () => {
         <ExerciseFive />
       </Route>
       <Route path="/exercise-six">
-        <ExerciseSix />
+        <ExerciseSix storage={sessionStorage} />
       </Route>
       <Route path="/exercise-seven">
         <ExerciseSeven />
       </Route>
       <Route path="/exercise-eight">
         <ExerciseEight percentage={70} />
-        </Route>
-
-  
-
+      </Route>
+      <Route path="/exercise-nine">
+        <ExerciseNine />
+      </Route>
+      <Route path="/exercise-ten">
+        <ExerciseTen />
+      </Route>
+      <Route path="/exercise-eleven">
+        <ExerciseEleven />
+      </Route>
     </BrowserRouter>
 
   )
